% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tda.R
\name{fcp_stats}
\alias{fcp_stats}
\title{Fréchet Change Point Detection Stats}
\usage{
fcp_stats(dist_mat_sq)
}
\arguments{
\item{dist_mat_sq}{nxn matrix of squared distances}
}
\description{
Fréchet Change Point Detection Stats
}
