% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet.R
\name{inco_var}
\alias{inco_var}
\title{A Fréchet variance statistic without need for calculating a Fréchet mean.}
\usage{
inco_var(dist_mat_sq)
}
\arguments{
\item{dist_mat_sq}{nxn matrix of squared distances}
}
\description{
A Fréchet variance statistic without need for calculating a Fréchet mean.
}
