% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{rmae_eval}
\alias{rmae_eval}
\title{Evaluate results of simulation by calculating the relative mean absolute error.}
\usage{
rmae_eval(results, samplers, estimators)
}
\arguments{
\item{results}{The results of a simulation.}

\item{samplers}{The table of samplers used in the simulation.}

\item{estimators}{The table of estimators used in the simulation.}
}
\value{
A table containing the relative mean absolute error (rmae) for each run.
}
\description{
Evaluate results of simulation by calculating the relative mean absolute error.
}
