% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Run Simulation}
\usage{
simulate(samplers, estimators, required_dists)
}
\arguments{
\item{samplers}{A tibble. Each row describes one distribution for a time
series of sets of points.}

\item{estimators}{A tibble. Each row describes an estimator of the change
point in the time series.}

\item{required_dists}{A tibble. Each row describes a distance of persistence
diagrams.}
}
\value{
A tibble. Each row describes the results of applying one estimator to
a time series created by one of the samplers.
}
\description{
Run Simulation
}
