% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_ts}
\alias{simulate_ts}
\title{Run a Simulation of a Time Series}
\usage{
simulate_ts(samplers, estimators, required_dists, n_cores = 1)
}
\arguments{
\item{samplers}{A tibble. Each row describes one distribution for a time
series of sets of points.}

\item{estimators}{A tibble. Each row describes an estimator of the change
point in the time series.}

\item{required_dists}{A tibble. Each row describes a distance of persistence
diagrams.}

\item{n_cores}{An integer. The number of cores to use. If greater 1, than R
package parallel is used to parallelize repetitions of experiments.}
}
\value{
A tibble. Each row describes the results of applying one estimator to
a time series created by one of the samplers.
}
\description{
Run a Simulation of a Time Series
}
