% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdaCpdSim-package.R
\docType{package}
\name{TdaCpdSim-package}
\alias{TdaCpdSim}
\alias{TdaCpdSim-package}
\title{TdaCpdSim: Evaluate Change Point Detection via Topological Data Analysis by Simulations}
\description{
With this package you can simulate a time series of point configurations, create persistence diagrams from this, apply TDA-based change point detection algorithms, and evaluate the simulation results.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de}

}
\keyword{internal}
